requirements.txt
aiohttp
async-timeout
attrs
chardet
discord.py
idna
multidict
PyMySQL
yarl
APScheduler
chardet
emoji
pytz
six
tzlocal
websockets